Should exit with a non-zero error code on fatal error
Currently there's a bunch of different conditions in `func main()` which cause an immediate `return` (exit) because of an error that was encountered. On golang, returning the main function causes the program to exit status code 0 (success). Rather than indicate success, failures should exit with a non-zero code so that process managers can recognise it as a failure and do something appropriate (e.g. restart if the error condition is transient).

Signed-off-by: Andrew Rowson <github@growse.com>